Implementa importador de indicadores a partir da planilha Excel.

+++ database/migrations/2018_09_25_141424_create_data_table.php
@@ -17,7 +17,7 @@ class CreateDataTable extends Migration
             $table->increments('id');
 			$table->morphs('datable');
 			$table->date('date');
-			$table->integer('value');
+			$table->decimal('value', 15, 4);
             $table->timestamps();
         });
     }
